Refactor analytics retrieval to global-firestore-data
